lab_2/w2-test-tools-master-3b907520ee6b249418a8fa5fe0f612bab4c9b808/src/main/java/uj/jwzp/w2/e3/SellingService.java
package uj.jwzp.w2.e3;

import uj.jwzp.w2.e3.external.PersistenceLayer;

import java.math.BigDecimal;

public class SellingService {
    private final PersistenceLayer persistenceLayer;
    final CustomerMoneyService moneyService;
    private final DiscountsService discountsService;

    public SellingService(PersistenceLayer persistenceLayer, DiscountsService discountsService) {
        this.persistenceLayer = persistenceLayer;
        this.persistenceLayer.loadDiscountConfiguration();
        this.moneyService = new CustomerMoneyService(this.persistenceLayer);
        this.discountsService = discountsService;
    }

    public boolean sell(Item item, int quantity, Customer customer) {
        BigDecimal price = item.getPrice().subtract(discountsService.getDiscountForItem(item, customer)).multiply(BigDecimal.valueOf(quantity));
        if (discountsService.isWeekendPromotion() && price.compareTo(BigDecimal.valueOf(5)) > 0) {
            price = price.subtract(BigDecimal.valueOf(3));
        }
        boolean sold = moneyService.pay(customer, price);
        if (sold) {
            return persistenceLayer.saveTransaction(customer, item, quantity);
        } else {
            return sold;
        }
    }

}
